Remove code that duplicated active_support
