+++ lib/dog_collar/logging/logger.rb
@@ -2,17 +2,12 @@
 
 require 'dog_collar/logging/formatters/json'
 require 'dog_collar/logging/formatters/pretty'
+require 'dog_collar/logging/log_methods'
 
 module DogCollar
   module Logging
     class Logger < Logger
-      LOG_SEV = {
-        debug: Logger::DEBUG,
-        info: Logger::INFO,
-        warn: Logger::WARN,
-        error: Logger::ERROR,
-        fatal: Logger::FATAL
-      }.freeze
+      include LogMethods
 
       attr_accessor :formatter
 
@@ -48,36 +43,11 @@ module DogCollar
         self.formatter = kwargs.fetch(:formatter, default_formatter)
       end
 
-      LOG_SEV.each do |method_name, severity|
-        define_method(method_name) do |message = nil, **meta, &block|
-          add(severity, message, **meta, &block)
-        end
-
-        # Allow severity checks to work even when level is overriden in older
-        # versions of Ruby (< 2.7.0).
-        define_method("#{method_name}?") do
-          level <= severity
-        end
-      end
-
       def with(**meta)
         child = dup
         child.before_log { meta }
         child
       end
-
-      def add(severity, message = nil, **meta, &block)
-        severity ||= ::Logger::UNKNOWN
-        return true if @logdev.nil? || severity < level
-
-        meta = execute_before_log_hooks.update(meta)
-        message = evaluate_log_block(message, meta, &block)
-
-        @logdev.write(format_message(severity, Time.now, progname, message, meta))
-        true
-      end
-
-      alias log add
 
       private
