+++ spec/contrib/rails/logger_spec.rb
@@ -7,6 +7,13 @@ describe DogCollar::Contrib::Rails::Logger do
 
   subject { described_class.new(io, service: 'foo') }
 
+  describe '#add' do
+    it 'should support log tags' do
+      subject.info('hi', foo: 'bar')
+      expect(JSON.parse(io.string)).to include('message' => 'hi', 'foo' => 'bar')
+    end
+  end
+
   describe '#silence' do
     context 'when silencer is true' do
       before do
@@ -34,54 +41,6 @@ describe DogCollar::Contrib::Rails::Logger do
 
         expect(io.string).not_to be_empty
       end
-    end
-  end
-
-  describe '#log_at' do
-    it 'yields' do
-      expect { |b| subject.log_at(:error, &b) }.to yield_control
-    end
-
-    it 'changes the log level' do
-      expect(subject.level).to be(Logger::Severity::DEBUG)
-
-      subject.log_at(:error) do
-        expect(subject.level).to be(Logger::Severity::ERROR)
-      end
-
-      expect(subject.level).to be(Logger::Severity::DEBUG)
-    end
-
-    it 'changes back even when the block raises an exception' do
-      expect {
-        subject.log_at(:error) { raise 'boom!' }
-      }.to raise_error(StandardError)
-
-      expect(subject.level).to be(Logger::Severity::DEBUG)
-    end
-
-    it 'is thread-safe' do
-      inside_second_log_at = Notifier.new
-      log_at_started = Notifier.new
-
-      t = Thread.new do
-        subject.log_at(:error) do
-          log_at_started.notify
-          inside_second_log_at.wait_for_notification
-          expect(subject.level).to eq(Logger::Severity::ERROR)
-        end
-      end
-
-      t.abort_on_exception = true
-      t.report_on_exception = false if t.respond_to?(:report_on_exception)
-      log_at_started.wait_for_notification
-
-      subject.log_at(:warn) do
-        inside_second_log_at.notify
-        expect(subject.level).to eq(Logger::Severity::WARN)
-      end
-
-      t.join
     end
   end
 end
